conda_requirements.txt
# packages in environment at /home/james/anaconda3/envs/gridworld:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main  
astroid                   2.3.3                    py37_0  
attrs                     19.3.0                   pypi_0    pypi
blas                      1.0                         mkl  
ca-certificates           2020.1.1                      0  
certifi                   2019.11.28               py37_0  
cffi                      1.13.2           py37h2e261b9_0  
chardet                   3.0.4                    pypi_0    pypi
click                     7.0                      pypi_0    pypi
cloudpickle               1.2.2                    pypi_0    pypi
colorama                  0.4.3                    pypi_0    pypi
cudatoolkit               10.1.243             h6bb024c_0  
decorator                 4.4.1                    pypi_0    pypi
filelock                  3.0.12                   pypi_0    pypi
freetype                  2.9.1                h8a8886c_1  
funcsigs                  1.0.2                    pypi_0    pypi
future                    0.18.2                   pypi_0    pypi
gym                       0.15.4                   pypi_0    pypi
idna                      2.8                      pypi_0    pypi
imageio                   2.6.1                    pypi_0    pypi
imageio-ffmpeg            0.3.0                    pypi_0    pypi
importlib-metadata        1.3.0                    pypi_0    pypi
intel-openmp              2019.4                      243  
isort                     4.3.21                   py37_0  
joblib                    0.14.1                     py_0  
jpeg                      9b                   h024ee3a_2  
jsonschema                3.2.0                    pypi_0    pypi
lazy-object-proxy         1.4.3            py37h7b6447c_0  
libedit                   3.1.20181209         hc058e9b_0  
libffi                    3.2.1                hd88cf55_4  
libgcc-ng                 9.1.0                hdf63c60_0  
libgfortran-ng            7.3.0                hdf63c60_0  
libpng                    1.6.37               hbc83047_0  
libprotobuf               3.11.1               h8b12597_0    conda-forge
libstdcxx-ng              9.1.0                hdf63c60_0  
libtiff                   4.1.0                h2733197_0  
mccabe                    0.6.1                    py37_1  
mkl                       2019.4                      243  
mkl-service               2.3.0            py37he904b0f_0  
mkl_fft                   1.0.15           py37ha843d7b_0  
mkl_random                1.1.0            py37hd6b4f25_0  
more-itertools            8.0.2                    pypi_0    pypi
moviepy                   1.0.1                    pypi_0    pypi
ncurses                   6.1                  he6710b0_1  
ninja                     1.9.0            py37hfd86e86_0  
numpy                     1.17.4           py37hc1035e2_0  
numpy-base                1.17.4           py37hde5b4d6_0  
olefile                   0.46                     py37_0  
opencv-python             4.1.2.30                 pypi_0    pypi
openssl                   1.1.1d               h7b6447c_4  
packaging                 19.2                     pypi_0    pypi
pillow                    6.2.1            py37h34e0f95_0  
pip                       19.3.1                   py37_0  
pluggy                    0.13.1                   pypi_0    pypi
proglog                   0.1.9                    pypi_0    pypi
protobuf                  3.11.1           py37he1b5a44_0    conda-forge
psutil                    5.6.7                    pypi_0    pypi
py                        1.8.0                    pypi_0    pypi
py-spy                    0.3.1                    pypi_0    pypi
pycparser                 2.19                     py37_0  
pyglet                    1.3.2                    pypi_0    pypi
pylint                    2.4.4                    py37_0  
pyparsing                 2.4.5                    pypi_0    pypi
pyqt5                     5.13.2                   pypi_0    pypi
pyqt5-sip                 12.7.0                   pypi_0    pypi
pyrsistent                0.15.6                   pypi_0    pypi
pytest                    5.3.1                    pypi_0    pypi
python                    3.7.5                h0371630_0  
pytorch                   1.3.1           py3.7_cuda10.1.243_cudnn7.6.3_0    pytorch
pyyaml                    5.2                      pypi_0    pypi
ray                       0.7.6                    pypi_0    pypi
readline                  7.0                  h7b6447c_5  
redis                     3.3.11                   pypi_0    pypi
requests                  2.22.0                   pypi_0    pypi
rope                      0.16.0                   pypi_0    pypi
scikit-learn              0.22.1           py37hd81dba3_0  
scipy                     1.3.3                    pypi_0    pypi
setproctitle              1.1.10                   pypi_0    pypi
setuptools                42.0.1                   py37_0  
six                       1.13.0                   py37_0  
sqlite                    3.30.1               h7b6447c_0  
tabulate                  0.8.3                    py37_0  
tensorboardx              1.9                        py_0    conda-forge
tk                        8.6.8                hbc83047_0  
torchvision               0.4.2                py37_cu101    pytorch
tqdm                      4.40.2                   pypi_0    pypi
urllib3                   1.25.7                   pypi_0    pypi
wcwidth                   0.1.7                    pypi_0    pypi
wheel                     0.33.6                   py37_0  
wrapt                     1.11.2           py37h7b6447c_0  
xz                        5.2.4                h14c3975_4  
zipp                      0.6.0                    pypi_0    pypi
zlib                      1.2.11               h7b6447c_3  
zstd                      1.3.7                h0b5b093_0  
